carga de layout
carga de layout

+++ app/website/controllers/apiConciliacion.js
@@ -249,9 +249,9 @@ ApiConciliacion.prototype.get_creaConciliacionDetalle = function(req, res, next)
     { name: 'movimientoID', value: req.query.movimientoID, type: self.model.types.INT },
     { name: 'CCP_IDDOCTO', value: req.query.CCP_IDDOCTO, type: self.model.types.STRING },
     { name: 'VIN', value: req.query.VIN, type: self.model.types.STRING },
-    { name: 'interesGrupoAndrade', value: req.query.interesGrupoAndrade, type: self.model.types.INT },
-    { name: 'interesFinanciera', value: req.query.interesFinanciera, type: self.model.types.INT },
-    { name: 'interesAjuste', value: req.query.interesAjuste, type: self.model.types.INT },
+    { name: 'interesGrupoAndrade', value: req.query.interesGrupoAndrade, type: self.model.types.DECIMAL },
+    { name: 'interesFinanciera', value: req.query.interesFinanciera, type: self.model.types.DECIMAL },
+    { name: 'interesAjuste', value: req.query.interesAjuste, type: self.model.types.DECIMAL },
     { name: 'situacion', value: req.query.situacion, type: self.model.types.INT },
     { name: 'checked', value: req.query.checked, type: self.model.types.INT }];
 
@@ -270,9 +270,9 @@ ApiConciliacion.prototype.get_guardaConciliacionDetalle = function(req, res, nex
     { name: 'movimientoID', value: req.query.movimientoID, type: self.model.types.INT },
     { name: 'CCP_IDDOCTO', value: req.query.CCP_IDDOCTO, type: self.model.types.STRING },
     { name: 'VIN', value: req.query.VIN, type: self.model.types.STRING },
-    { name: 'interesGrupoAndrade', value: req.query.interesGrupoAndrade, type: self.model.types.INT },
-    { name: 'interesFinanciera', value: req.query.interesFinanciera, type: self.model.types.INT },
-    { name: 'interesAjuste', value: req.query.interesAjuste, type: self.model.types.INT },
+    { name: 'interesGrupoAndrade', value: req.query.interesGrupoAndrade, type: self.model.types.DECIMAL },
+    { name: 'interesFinanciera', value: req.query.interesFinanciera, type: self.model.types.DECIMAL },
+    { name: 'interesAjuste', value: req.query.interesAjuste, type: self.model.types.DECIMAL },
     { name: 'situacion', value: req.query.situacion, type: self.model.types.INT },
     { name: 'checked', value: req.query.checked, type: self.model.types.INT }];
